% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fetch.R
\name{cell_dist}
\alias{cell_dist}
\title{Calculate distances by cell counts}
\usage{
cell_dist(
  indices_df,
  cell_res = 2000,
  max_dist = 1e+05,
  func = "mean",
  na_remove = TRUE
)
}
\arguments{
\item{indices_df}{dataframe of indices}

\item{cell_res}{numeric cell resolution. Default is 2000.}

\item{max_dist}{numeric max distance to calculate fetch distance. Default is 100000 (100km).}

\item{func}{character R function to apply to distances from each cell. Either mean, min, max, or sum.}

\item{na_remove}{logical. argument passed to na.rm argument in R func. Default is TRUE, remove NAs}
}
\value{
dataframe of summarized cell distances
}
\description{
Calculate distances by cell counts
}
