% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fetch.R
\name{check_cols}
\alias{check_cols}
\title{Check boundaries of matrix column}
\usage{
check_cols(m, row_index, col_index)
}
\arguments{
\item{m}{matrix of interest}

\item{row_index}{numeric row index}

\item{col_index}{numeric col index}
}
\value{
list of cell indexes
}
\description{
Check boundaries of matrix column
}
