% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fetch.R
\name{check_length}
\alias{check_length}
\title{Check length of consecutive non-zero cells in vector}
\usage{
check_length(vect, max_count, verbose = FALSE)
}
\arguments{
\item{vect}{numeric vector of 1s and 0s (can include NAs)}

\item{max_count}{numeric, indicating max number of pixels to count within maximum distance}

\item{verbose}{logical, whether messages should be printed. Default is FALSE, no messages print.}
}
\value{
Count of consecutive non-zero cells from a matrix cell, extending out in a given direction
}
\description{
Check length of consecutive non-zero cells in vector
}
