% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fetch.R
\name{check_zeros}
\alias{check_zeros}
\title{Check run length encoding output for cell run of zeros}
\usage{
check_zeros(rle_vect, verbose = FALSE)
}
\arguments{
\item{rle_vect}{rle object}

\item{verbose}{logical, whether messages should be printed. Default is FALSE, no messages print.}
}
\value{
Count of consecutive non-zero cells from a matrix cell, extending out in a given direction
}
\description{
Check run length encoding output for cell run of zeros
}
