% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fetch.R
\name{diag_summary}
\alias{diag_summary}
\title{Summarize lengths of sequential raster cells diagonals}
\usage{
diag_summary(
  m,
  indices_df,
  func = "mean",
  max_count,
  ncores = 2,
  verbose = FALSE
)
}
\arguments{
\item{m}{matrix of interest}

\item{indices_df}{dataframe of matrix indices from which to calculate lengths from}

\item{func}{character R function to summarize surrounding vertical/horizontal pixels. One of "mean", "min", "max", or "sum". Default is "mean"}

\item{max_count}{numeric, maximum cell count. Equal to maximum distance/cell resolution}

\item{ncores}{numeric, number of cores to use for parallel processing. Default is 2.}

\item{verbose}{logical, whether to print messages or not. Default is FALSE.}
}
\value{
list of numeric vectors of diagonal distances
}
\description{
Internal function for calculating distance along diagonals for specified indices of a matrix.
}
