% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{fetch-class}
\alias{fetch-class}
\title{Fetch Class}
\description{
Fetch Class
}
\section{Slots}{

\describe{
\item{\code{location_lat}}{latitude of the location of interest.}

\item{\code{location_lon}}{longitude of the location of interest.}

\item{\code{subset_coast}}{a subset of the NZ coast shapefile to increase efficiency.}

\item{\code{subset_island}}{a subset of the NZ offshore islands shapefile to increase
                    efficiency.}
}}

