% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R, R/fetch_class.R
\docType{methods}
\name{fetch}
\alias{fetch}
\alias{show,Fetch-method}
\title{Calculate Fetch for a New Zealand Coastal Marine Location}
\usage{
fetch(lat, lon, max_dist = 300, n_bearings = 9, quiet = FALSE, accuracy,
  degree_int)

\S4method{show}{Fetch}(object)
}
\arguments{
\item{lat}{latitude in decimal degrees.}

\item{lon}{longitude in decimal degrees.}

\item{max_dist}{maximum distance in kilometers (default 300).}

\item{n_bearings}{the number of bearings per quadrant. A value of 9 (default)
gives a total of 36 bearings i.e. one at every 10 degrees.}

\item{quiet}{suppress diagnostic messages (default \code{FALSE}).}

\item{accuracy}{Deprecated.}

\item{degree_int}{Deprecated.}

\item{object}{a \code{\link{Fetch}} object.}
}
\value{
Returns a \code{\link{Fetch}} object.
}
\description{
Fetch is the unobstructed length of water over which wind can blow, and is
commonly used as a measure of exposure to wind and waves at coastal sites. The 
\code{fetch} function calculates the fetch for a New Zealand coastal location 
for a number of bearings to quantify the fetch in 
various compass directions.
}
\examples{
# Calculate fetch for Kawau Bay
kawau_bay = fetch(-36.4, 174.8)

# Show the distances for each direction and the resultant locations
kawau_bay

# Summarise the information
summary(kawau_bay)

# Plot the vectors
plot(kawau_bay)

\dontrun{
# ---- Output to KML/KMZ
# Install and load plotKML library
if (!require(plotKML))
  install.packages("plotKML")
library(plotKML)

Create some labels indicating the fetch vector directions
labs = sapply(slot(kawau_bay, "lines"), slot, "ID")

# Save 'kawau_bay.kml' to the current directory
kml(kawau_bay, labels = labs)
}
}
\seealso{
\code{\link{fetchR}}
}

