% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_class.R
\name{fetchR-deprecated}
\alias{fetchR-deprecated}
\alias{save_kml}
\title{Deprecated Functions in \pkg{fetchR}}
\usage{
save_kml()
}
\description{
\code{save_kml} has been deprecated. Instead use the \code{kml} function in
the \pkg{plotKML} package.
}
\examples{
\dontrun{
kawau_bay = fetch(-36.4, 174.8)

# Install plotKML if required
if (!require(plotKML))
  install.packages("plotKML")
  
library(plotKML)

# Create some labels indicating the fetch vector directions
labs = sapply(slot(kawau_bay, "lines"), slot, "ID")

# Save 'kawau_bay.kml' to the current directory
kml(kawau_bay, labels = labs)
}

}

