% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_class.R
\docType{package}
\name{fetchR}
\alias{fetchR}
\alias{fetchR-package}
\title{Calculate Wind Exposure with the \pkg{fetchR} Package}
\description{
The \pkg{fetchR} package provides an objective calculation of wind fetch.
}
\details{
Fetch is an important measurement in coastal applications. It
provides a measurement for the unobstructed length of water that wind from a
certain direction can blow over. The higher the wind fetch from a certain 
direction, the more energy is imparted onto the surface of the 
water resulting in a larger sea state. Therefore, the larger
the fetch, the larger the exposure to wind and the more likely the
site experiences larger sea states.

The fetch length from all directions (and from each quadrant) can therefore 
be averaged to provide an overall indication of the location's exposure to 
wind, and from which directions and/or quadrants. The \pkg{fetchR} package 
calculates the lengths of wind fetch vectors from all directions, at any 
given location(s) on Earth, and provides summaries, visualisations and KML 
files along with the raw data.
}
\seealso{
\code{\link{fetch}} for reproducible examples for calculating fetch, 
         along with summarising, visualising and exporting these calculations.
}
\keyword{package}

