% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_landwater.R
\name{filepath_error_msg}
\alias{filepath_error_msg}
\title{Internal filepath error handling message}
\usage{
filepath_error_msg()
}
\value{
character error message for invalid file paths
}
\description{
Internal filepath error handling message
}
