% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_landwater.R
\name{fix_grid}
\alias{fix_grid}
\title{Align irregular grid to fit a regular grid with a specified resolution}
\usage{
fix_grid(r, res = 2000, verbose = FALSE)
}
\arguments{
\item{r}{terra SpatRaster or raster RasterLayer object in a projected CRS.}

\item{res}{numeric resolution of the aligned output raster. Default is 2000.}

\item{verbose}{logical, whether messages should be printed. Default is FALSE, so no messages will print.}
}
\value{
SpatRaster on a regular grid at a specified resolution
}
\description{
Internal utils function for generating landwater rasters
}
