% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fetch.R
\name{get_fetch}
\alias{get_fetch}
\title{Generate a raster of mean fetch lengths from a binary land-water raster.}
\usage{
get_fetch(r = NULL, max_dist = 1e+05, ncores = 2, verbose = TRUE)
}
\arguments{
\item{r}{SpatRaster, RasterLayer or file path (character) to binary land type raster TIF. Land and water values only.}

\item{max_dist}{numeric maximum distance (meters) to calculate fetch values out to. Default is 100000m (100km)}

\item{ncores}{numeric indicating how many cores to use during parallel processing. Default is 2.}

\item{verbose}{logical, whether messages should be printed. Default is FALSE, no messages print.}
}
\value{
SpatRaster with the mean fetch distance from land calculated for every water pixel.
}
\description{
Given a binary land and water raster, this function will calculate the mean distance a water cell is away from land. The fetch distance is the mean distance between the given water cell and the next land cell in 8 directions (horizontally, vertically, and diagonally). Returns a raster of the same CRS, resolution, and extent as the input land water raster.
}
