% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_landwater.R
\name{get_land}
\alias{get_land}
\title{Make a land raster from binary land type raster}
\usage{
get_land(r, land_value = 0, out_value = 2, as_polygon = FALSE, as_sf = FALSE)
}
\arguments{
\item{r}{SpatRast of binary land type raster. Land and water values only.}

\item{land_value}{numeric indicating what value land is represented by in raster. Default is 0 to conform to get_landwater() output}

\item{out_value}{numeric indicating what value land should be represented by in output raster. Default is 2}

\item{as_polygon}{logical, should polygon be returned instead of a raster. Default is FALSE, returns a terra rast object}

\item{as_sf}{logical, should polygon be returned as an SF object. Default is FALSE, returns a terra vect object}
}
\value{
SpatRaster with only land cells having a value, all other cells are NA
}
\description{
returns a SpatRaster of only land cells. Land cells have a value of 2 by default, all other cells are NA. Fits the HSI raster specifications.
}
