% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_landwater.R
\name{get_landwater}
\alias{get_landwater}
\title{Make a binary land water raster from multiple land types raster.}
\usage{
get_landwater(r, water_value = NA, res = 2000, verbose = TRUE)
}
\arguments{
\item{r}{SpatRaster, RasterLayer, SpatVect or a filepath (character) to raster data. Data must be on a projected CRS.}

\item{water_value}{numeric, value of water cells in input raster. Default is NA.}

\item{res}{numeric resolution for template raster. Default is 2000, generates a raster with a grid cell resolution of 2000x2000.}

\item{verbose}{logical, whether messages should be printed. Default is TRUE, no messages print.}
}
\value{
SpatRaster, a binary land water raster with land cells set to 1 and water cells set to 0.
}
\description{
returns a binary land water raster with land cells set to 1 and water cells set to 0. A water value should be specified to indicate which pixel value should represent water in the final raster. The rest of the pixels will indicate land cells.
}
