% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_landwater.R
\name{get_water}
\alias{get_water}
\title{Make a water raster from binary land type raster}
\usage{
get_water(r, water_value = 1, out_value = 1, as_polygon = FALSE, as_sf = FALSE)
}
\arguments{
\item{r}{SpatRast of binary land type raster. Land and water values only.}

\item{water_value}{numeric indicating what value water is represented by in raster. Default is 1 to conform to get_landwater() output}

\item{out_value}{numeric indicating what value land should be represented by in output raster. Default is 1}

\item{as_polygon}{logical, should polygon be returned instead of a raster. Default is FALSE, returns a terra rast object}

\item{as_sf}{logical, should polygon be returned as an SF object. Default is FALSE, returns a terra vect object}
}
\value{
SpatRaster with only land cells having a value, all other cells are NA
}
\description{
returns a SpatRaster of only water cells. Water cells have a value of 1 by default, all other cells are NA. Fits the HSI raster specifications.
}
