% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_data.R
\docType{data}
\name{land}
\alias{land}
\title{Land raster}
\format{
An object of class \code{PackedSpatRaster} of length 1.
}
\usage{
land
}
\description{
Example land raster. Values of 1 indicate land and all other values are NA. Implied that NA values are water cells
}
\examples{
\dontrun{
 fetchr::land
}

}
\keyword{datasets}
