% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_vect_data.R
\docType{data}
\name{land_vect}
\alias{land_vect}
\title{Land vector polygon}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 2 columns.
}
\usage{
land_vect
}
\description{
Example land vector polygon
}
\examples{
\dontrun{
 fetchr::land_vect
}

}
\keyword{datasets}
