% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landwater_data.R
\docType{data}
\name{landwater}
\alias{landwater}
\title{Binary landwater raster}
\format{
An object of class \code{RasterLayer} of dimension 189 x 183 x 1.
}
\usage{
landwater
}
\description{
Example binary landwater raster. Values of 1 indicate land and values of 0 indicate water.
}
\examples{
\dontrun{
 fetchr::landwater
}

}
\keyword{datasets}
