% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_landwater.R
\name{make_grid}
\alias{make_grid}
\title{Create grid template}
\usage{
make_grid(r, res = 2000)
}
\arguments{
\item{r}{SpatRaster, SpatVector or SF object with a Projected Coordinate System CRS.}

\item{res}{numeric resolution for template raster. Default is 2000, generates a template raster with a grid cell resolution of 2000x2000.}
}
\value{
SpatRaster of template grid
}
\description{
returns a template SpatRaster. All cell values are 1 so the template raster is not actually "empty
}
