% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,fetch,missing-method}
\title{Plot a Fetch Object}
\usage{
\S4method{plot}{fetch,missing}(x, y, ...)
}
\arguments{
\item{x}{\code{fetch} object as returned by \code{\link{fetch}}.}

\item{y}{missing (not used).}

\item{...}{further arguments passed to \code{\link{segments}}}
}
\description{
This is the default plot method for a \code{\link{fetch}} object. A map is
plotted with the area determined by the length of the direction vectors. The
vectors are plotted as lines originating from the location's coordinates.
}

