% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_landwater.R
\name{prep_filepath}
\alias{prep_filepath}
\title{Prepare filepath vector/raster data for fetch calculations}
\usage{
prep_filepath(r, water_value = NA, res = 2000, verbose = FALSE)
}
\arguments{
\item{r}{character, file path to a shp, gpkg, geojson, or tif file in a projected CRS}

\item{water_value}{numeric, value indicating the value of the water cells within the provided tif file. If a shp, gpkg, or geojson is given, this argument is ignored. Default is NA, such that NA cells are treated as water cells and all other cells are land cells.}

\item{res}{numeric resolution of the aligned output raster. Default is 2000.}

\item{verbose}{logical, whether messages should be printed. Default is FALSE, no messages print.}
}
\value{
SpatRaster, binary landwater raster with land values of 0 and water values of 1
}
\description{
Internal function for checking and aligning raster data to specifications for calculating fetch using get_fetch
}
