% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_landwater.R
\name{prep_poly}
\alias{prep_poly}
\title{Prepare polygon vector data for fetch calculations}
\usage{
prep_poly(r, res = 2000, verbose = FALSE)
}
\arguments{
\item{r}{terra SpatVector POLYGON/MULTIPOLYGON or sf POLYGON/MULTIPOLYGON object in a projected CRS}

\item{res}{numeric resolution of the aligned output raster. Default is 2000.}

\item{verbose}{logical, whether messages should be printed. Default is FALSE, no messages print.}
}
\value{
SpatRaster, binary landwater raster with land values of 0 and water values of 1
}
\description{
Internal function for checking and aligning polygon data to specifications for calculating fetch using get_fetch
}
