% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_landwater.R
\name{prep_raster}
\alias{prep_raster}
\title{Prepare gridded raster data for fetch calculations}
\usage{
prep_raster(r, water_value = NA, verbose = FALSE)
}
\arguments{
\item{r}{terra SpatRaster or raster RasterLayer object in a projected CRS.}

\item{water_value}{numeric, value indicating the value of the water cells within the input raster. If a sf/terra polygon is given, this argument is ignored. Default is NA, such that NA cells are treated as water cells and all other cells are land cells.}

\item{verbose}{logical, whether messages should be printed. Default is FALSE, no messages print.}
}
\value{
SpatRaster, binary landwater raster with land values of 0 and water values of 1
}
\description{
Internal function for checking and aligning raster data to specifications for calculating fetch using get_fetch
}
