% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_class.R
\docType{methods}
\name{proj4string,Fetch-method}
\alias{proj4string,Fetch-method}
\title{Retrieve projection attributes for Fetch objects}
\usage{
\S4method{proj4string}{Fetch}(obj)
}
\arguments{
\item{obj}{\code{\link{Fetch}} object}
}
\description{
Retrieve projection attributes for \code{\link{Fetch}} objects.
}

