% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run fetchR example}
\usage{
runExample()
}
\description{
Launch a Shiny application of \pkg{fetchR} that incorporates all of the
functionality available within the package.
}
\details{
The application is also
\href{https://blasee.shinyapps.io/fetchR_shiny/}{available online}.
}
\examples{
## Only run this application in interactive R sessions
if (interactive())
  runExample()

}
\seealso{
For help on using the \pkg{fetchR} web application refer to the
\href{https://github.com/blasee/fetchR/blob/master/README_shiny.md#calculate-wind-fetch-using-the-shiny-application}{README}.
}

