% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{save_kml}
\alias{save_kml}
\alias{save_kml,fetch,ANY-method}
\title{Save a KML file of the direction vectors}
\usage{
\S4method{save_kml}{fetch}(object, file_name)
}
\description{
The \code{save_kml} function saves a KML file for use in Google Earth or
other KML-friendly software.
}
\details{
This function is useful for checking that no vectors pass through any
subtidal reefs or other potentially fetch-limiting structures that aren't
defined in the original shape file.
}

