% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fetch.R
\name{side_count}
\alias{side_count}
\title{Count number of consequetive cells around a matrix cell to the up/down/left/right of a matrix cell}
\usage{
side_count(
  m,
  row_index = NULL,
  col_index = NULL,
  cell_res = 2000,
  max_dist = 1e+05,
  func = "mean",
  as_df = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{m}{matrix of interest}

\item{row_index}{numeric row number of matrix cell to calculate cell count going up/down/left/right of the specified cell}

\item{col_index}{numeric column number of matrix cell to calculate cell count going up/down/left/right of the specified cell}

\item{cell_res}{numeric cell resolution. Default is 2000.}

\item{max_dist}{numeric maximum distance to calculate fetch to. Default is 100000m (100km).}

\item{func}{character. R function, either, mean, min, max, or sum. Default is mean calculates mean number of cells around cell}

\item{as_df}{logical. whether to return a dataframe or not. Default is FALSE, returns numeric vector of mean cell count around cell.}

\item{verbose}{logical, whether to print messages to console. Default is FALSE, no messages print}
}
\value{
numeric indicating the average cell count in all directions or a dataframe with consequetive cell counts (non-zero cells) up/down/left/right of the specified cell in a matrix
}
\description{
Count number of consequetive cells around a matrix cell to the up/down/left/right of a matrix cell
}
