% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fetch.R
\name{side_summary}
\alias{side_summary}
\title{Summarize lengths of sequential raster cells horizontalally and vertically}
\usage{
side_summary(
  m,
  indices_df,
  cell_res = 2000,
  max_dist = 1e+05,
  func = "mean",
  cell_count = NULL,
  ncores = 2,
  verbose = FALSE
)
}
\arguments{
\item{m}{matrix of interest}

\item{indices_df}{dataframe of matrix indices from which to calculate lengths from}

\item{cell_res}{numeric, resolution of raster cells. Default is 2000}

\item{max_dist}{numeric, maximum distance to calculate fetch out to. in meters. Default is 20000m}

\item{func}{character R function to summarize surrounding vertical/horizontal pixels. One of "mean", "min", "max", or "sum". Default is "mean"}

\item{ncores}{numeric, number of cores to use for parallel processing. Default is 2.}

\item{verbose}{logical, whether to print messages or not. Default is FALSE.}
}
\value{
returns a list of summarized side lengths for every water cell, the same length as the number of rows in the input indices_df dataframe
}
\description{
Internal function for calculating distances horizonatlly and vertically for specified indices of a matrix.
}
