% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_class.R
\docType{methods}
\name{spTransform,Fetch,CRS-method}
\alias{spTransform,Fetch,CRS-method}
\alias{spTransform,Fetch,character-method}
\title{spTransform for map projection and datum transformation}
\usage{
\S4method{spTransform}{Fetch,CRS}(x, CRSobj)

\S4method{spTransform}{Fetch,character}(x, CRSobj)
}
\arguments{
\item{x}{\code{\link{Fetch}} object to be transformed}

\item{CRSobj}{object of class \code{\link[sp]{CRS}}, or of class character in
which case it is converted to \code{\link{CRS}}}

\item{...}{further arguments (ignored)}
}
\value{
\code{\link{Fetch}} object with coordinates transformed to the
        new coordinate system.
}
\description{
spTransform for map projection and datum transformation
}

