% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_class.R
\docType{methods}
\name{summary,Fetch-method}
\alias{summary,Fetch-method}
\title{Summarise a Fetch Object}
\usage{
\S4method{summary}{Fetch}(object)
}
\arguments{
\item{object}{\code{Fetch} object as returned by \code{\link{fetch}}.}
}
\description{
The \code{summary} function calculates the mean and median fetch of the 
location, along with the average fetch for the separate northerly, easterly, 
southerly and westerly quadrants. The mean fetch for the northerly component,
for example, averages over the fetch vectors between directions 315 (inclusive)
and 45 (exclusive), i.e. the fetch vectors within the interval [315, 45). 
The number of bearings per quadrant (\code{n_bearings}) is also returned to 
remind the user how many vectors were used for calculating the means for each 
quadrant.
}

