% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,fetch-method}
\title{Summarise a Fetch Object}
\usage{
\S4method{summary}{fetch}(object)
}
\arguments{
\item{object}{\code{fetch} object as returned by \code{\link{fetch}}.}
}
\description{
The \code{summary} function calculates the mean and median fetch and the most
exposed direction(s).
}

