% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_class.R
\docType{class}
\name{Fetch-class}
\alias{Fetch}
\alias{Fetch-class}
\title{Fetch Object}
\description{
Extends the \code{\link[sp]{SpatialLines}} class to include the subset of the 
coastal NZ map which can speed up plotting.
}
\section{Slots}{

\describe{
\item{\code{location_lat}}{latitude of the location of interest.}

\item{\code{location_lon}}{longitude of the location of interest.}

\item{\code{subset_map}}{a subset of the coastal NZ shapefile to increase efficiency when plotting.}
}}

