% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_class.R
\docType{class}
\name{Fetch}
\alias{Fetch}
\title{Fetch Class}
\description{
Class to hold Fetch objects.
}
\details{
A \code{Fetch} object is essentially a list of \code{\link[sp]{SpatialLines}}
objects.
}
\section{Slots}{

\describe{
\item{\code{names}}{character vector containing the names for each location.}

\item{\code{max_dist}}{numeric vector of length 1 containing the maximum distance
a fetch vector is allowed.}
}}
\note{
Fetch objects should only be created using the \code{\link{fetch}}
function.
}
\section{Extends}{

Class \code{"list"} directly, and class \code{"vector"}, by class \code{"list"}.
}

