% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fetch}
\alias{fetch}
\title{Calculate Fetch for a (NZ) Coastal Marine Location}
\usage{
fetch(lon, lat, max_dist = 300, accuracy = 0.1, n_angles = 10,
  plot = TRUE, zoom = max_dist/10, quiet = FALSE, ...)
}
\arguments{
\item{lon}{longitude in decimal degrees.}

\item{lat}{latitude in decimal degrees.}

\item{max_dist}{maximum distance (km) allowed for any given angle (default
300).}

\item{accuracy}{accuracy (km) of the fetch estimate (default 0.1km).}

\item{n_angles}{number of angles for the fetch calculation (equally spaced).}

\item{quiet}{\code{FALSE}. Suppress diagnostic messages.}

\item{plot}{\code{TRUE}. Create a png in the current working directory.}

\item{zoom}{number indicating the zoom for the plot.}

\item{...}{further arguments passed to \code{\link[grDevices]{png}}}
}
\value{
returns an invisible dataframe containing the resultant coordinates
        at the coastline, magnitude and direction.
}
\description{
Calculate average fetch for a New Zealand coastal location given in decimal
degrees. The accuracy can be as low as 1m (0.001km) but can be very
computationally expensive, especially for exposed sites. The number of angles
can be between 4 and 100, however large values can be very computationally
expensive and may not improve the accuracy of the estimate too much. When
land is hit, the distance to coast from the previous iteration is used i.e.
the points will always end up in the water. The zoom is only used for when
\code{plot} is \code{TRUE} --- the higher the value, the higher the zoom.
}
\seealso{
fetchR
}

