% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{fetchR}
\alias{fetchR}
\alias{fetchR-package}
\title{Calculate Fetch with \pkg{fetchR}}
\description{
The \pkg{fetchR} package provides an objective calculation of fetch
based on current conventional standards.
}
\details{
Fetch is an important measurement in coastal applications effectively
providing a proportional measurement of the length of water wind is blown
over. The higher the fetch the more energy is imparted to the surface of the
water resulting in a larger sea state
(\url{http://en.wikipedia.org/wiki/Fetch_(geography)}).

Fetch is currently calculated as the sum of distances from the marine
location to the shoreline for every 10 degrees (default) of the compass rose.
The only function required for the calculation of fetch at given latitude and
longitude coordinates is \code{\link{fetch}}.
}
\note{
The \pkg{fetchR} package is intended for New Zealand coastal
applications and only supports marine sites within the following bounding
box:

\tabular{lll}{
 \tab Min \tab Max \cr
Longitude \tab 167 \tab  178 \cr
Latitude  \tab -47 \tab  -35 \cr
}

Locations near the perimiter or outside this bounding box are either not
coastal or do not (yet?) have the high resolution coastal boundaries to
calculate the fetch.
}
\seealso{
\code{\link{fetch}}
}
\keyword{package}

