% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_class.R
\docType{methods}
\name{plot,Fetch,missing-method}
\alias{plot,Fetch,missing-method}
\title{Plot a Fetch Object}
\usage{
\S4method{plot}{Fetch,missing}(x, y, ...)
}
\arguments{
\item{x}{\code{fetch} object as returned by \code{\link{fetch}}.}

\item{y}{missing (not used).}

\item{...}{further arguments passed to \code{\link[sp]{plot,SpatialLines,missing-method}}.}
}
\description{
This is the default plot method for a \code{\link{Fetch}} object. The NZ
Transverse Mercator 2000 projection is used for the plot and the fetch 
vectors are plotted as lines originating from the location's coordinates.
}

