% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_class.R
\name{summary,Fetch-method}
\alias{summary,Fetch-method}
\alias{show,Fetch-method}
\title{Summarise a Fetch Object}
\usage{
\S4method{summary}{Fetch}(object)

\S4method{show}{Fetch}(object)
}
\arguments{
\item{object}{a \code{Fetch} object that has been returned by the
\code{\link{fetch}} function.}
}
\value{
The \code{summary} function returns a \code{\link{data.frame}}.
}
\description{
The \code{summary} function calculates the average fetch for the separate
northerly, easterly, southerly and westerly quadrants. For example, the mean
fetch for the northerly component averages over the fetch vectors with
directions between 315 (inclusive) and 45 (exclusive) degrees, i.e. the fetch
vectors within the interval [315, 45).
}
